+++ R/imports.R
@@ -1,5 +1,6 @@
 #' @importFrom stats as.ts AIC BIC logLik nobs fitted residuals coef time
-#' @importFrom zoo index
+#' @importFrom zoo as.zoo
 #' @importFrom xts as.xts
 #' @importFrom generics tidy augment glance
+#' @importFrom vctrs vec_ptype2 vec_cast
 NULL