+++ R/reexports.R
@@ -1,11 +1,3 @@
-#' @export
-#' @importFrom vctrs vec_ptype2
-vctrs::vec_ptype2
-
-#' @export
-#' @importFrom vctrs vec_cast
-vctrs::vec_cast
-
 #' @importFrom generics augment
 #' @export
 generics::augment