+++ R/pkg-segmented.R
@@ -6,7 +6,23 @@ as.seg_cpt.segmented <- function(object, ...) {
     x = as.ts(object),
     pkg = "segmented",
     base_class = class(object),
-    algorithm = "Segmented",
+    algorithm = "selgmented",
+    changepoints = changepoints(object),
+    seg_params = list(seg_params(object)),
+    model = model_name(object),
+    fitness = fitness(object)
+  )
+}
+
+#' @rdname as.segmenter
+#' @export
+#' 
+as.seg_cpt.stepmented <- function(object, ...) {
+  seg_cpt(
+    x = as.ts(object),
+    pkg = "segmented",
+    base_class = class(object),
+    algorithm = "stepmented",
     changepoints = changepoints(object),
     seg_params = list(seg_params(object)),
     model = model_name(object),
@@ -17,13 +33,23 @@ as.seg_cpt.segmented <- function(object, ...) {
 #' @rdname reexports
 #' @export
 as.ts.segmented <- function(x, ...) {
-  as.ts(x$model$obj)
+  stats::model.frame(x)[, 1] |>
+    as.ts()
 }
+
+#' @rdname reexports
+#' @export
+as.ts.stepmented <- function(x, ...) {
+  # where did the original data go???
+  fitted(x) + residuals(x) |>
+    as.ts()
+}
+
 
 #' @rdname changepoints
 #' @export
 #' @examples
-#' cpts <- segment(DataCPSim, method = "segmented")
+#' cpts <- segment(DataCPSim, method = "selgmented")
 #' changepoints(cpts$segmenter)
 #' 
 changepoints.segmented <- function(x, ...) {
@@ -32,25 +58,31 @@ changepoints.segmented <- function(x, ...) {
     as.integer()
 }
 
+#' @rdname changepoints
+#' @export
+#' @examples
+#' cpts <- segment(DataCPSim, method = "stepmented")
+#' changepoints(cpts$segmenter)
+#' 
+changepoints.stepmented <- function(x, ...) {
+  x$psi[[1, 1]] |>
+    round() |>
+    as.integer()
+}
+
 #' @rdname fitness
 #' @export
 #' @examples
 #' # Segment a time series using Segmented
-#' x <- segment(DataCPSim, method = "segmented")
+#' x <- segment(DataCPSim, method = "selgmented")
 #' 
 #' # Retrieve its fitness
 #' fitness(x)
 #' 
-fitness.segmented <- function(object, ...) {
-  out <- MDL(object)
-  names(out) <- "MDL"
+fitness.lm <- function(object, ...) {
+  out <- BIC(object)
+  names(out) <- "BIC"
   out
-}
-
-#' @rdname reexports
-#' @export
-nobs.segmented <- function(object, ...) {
-  length(as.ts(object))
 }
 
 #' @rdname model_name
@@ -59,15 +91,21 @@ model_name.segmented <- function(object, ...) {
   "trendshift"
 }
 
+#' @rdname model_name
+#' @export
+model_name.stepmented <- function(object, ...) {
+  "meanshift"
+}
+
 #' @rdname model_args
 #' @export
-model_args.segmented <- function(object, ...) {
+model_args.lm <- function(object, ...) {
   NULL
 }
 
 #' @rdname seg_params
 #' @export
-seg_params.segmented <- function(object, ...) {
+seg_params.lm <- function(object, ...) {
   list(
     sigma = sqrt(sum(residuals(object)^2))
   )