+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added support for models from the `strucchange` package.
 * Added group aesthetic to plots (@rpruim)
 * The dots (...) are passed more consistently
+* Moved `patchwork` to Suggests
 
 # tidychangepoint 1.0.2