+++ NEWS.md
@@ -5,6 +5,7 @@
 * The dots (...) are passed more consistently
 * Moved `patchwork` to Suggests
 * Moved `stats` reexports to imports
+* Moved `vctrs` reexports to imports
 * Moved `broom` reexports to `generics`
 
 # tidychangepoint 1.0.2